Cleanup: Remove dead / unreachable code (part 2 of 2)
* Remove also data structures and accessors related to
  selector code removed with previous commit.

+++ src/see/model/Range.java
@@ -277,7 +277,7 @@ public class Range implements Representation
    * Returns the minimally required bit size to code the total range.
    * @return The minimally required bit size in the range 0..32.
    */
-  public int getRequiredBitSize()
+  public byte getRequiredBitSize()
   {
     return (byte)Math.ceil(Math.log(size) / Math.log(2));
   }
